Add CheckoutPanel to Cart page
